Move fork/candidates into candidate
